function-source/requirements.txt
functions-framework==3.*
gcsfs
fsspec
google-cloud-storage
google-cloud-bigquery